Reduce frameskip to 2 frames max.
I originally thought the canvas copying code was a bottleneck,
but further profiling revealed some major inefficiencies in the core
project. With those fixed, performance is a lot better, and we can
dial this down as it wasn't very helpful. A little bit of it still
helps to keep the speed consistent though.