Add support for the second gamepad, enabling 2P mode!
Not all keys are made equal, and not all browsers are going to be
happy with certain keys being mapped, but that's for the users
to figure out. The default will be for P2 to be entirely unmapped
though, since trying to map two people onto one keyboard is
non-trivial, and will probably need to be adjusted for the users
taste and keyboard layout.