Begin work on an embeddable version of the HTML
The goal here is for this to be ideal to toss into an iframe, either
self-hosted or referencing the main hosted copy, with a link to a
cartridge to auto-load. I'd ideally like to maintain the control
configuration, but the main goal is for the screen to automatically
size itself to fill the container (maintaining aspect ratio) so the
end user only needs to drop it in and configure it minimally.