Switch to a 256x240 canvas overlay, and apply nearest neighbor
This is again a performance boost, as the browser is not applying
a fullscreen blur filter on the upscaled image. The resized source
has no effect on performance, but looks nicer withnearest-neighbor
as the pixelation now lines up with the NES screen, making the
pixelated aesthetic look intentional rather than buggy.

This change is about a 30% drop in CPU usage on my machine, which
is fairly substantial.