Fix glitchy lines on some versions of Firefox
Turns out canvas elements, due to CSS rounding, were sometimes
ending up on a subpixel boundary, and Firefox was getting a little
confused when applying a GPU context to the underlying canvas. This
doesn't fix the size mismatch, but it hides the resulting artifact
and seems to maintain equivalent performance.