Warn on an inactive audio context
Chroium-based browsers (incl. Edge, and possibly Safari) won't allow an audio worklet to begin processing until a user gesture activates the page. We've had that implemented for ages, but it's not obvious that this needs to be done for direct NSF links, so a warning message is in order.

Later I might want to make this prettier, and the embedded view will need a similar treatment.