Switch to the LQ audio path, and disable N163 multiplexing
This results in a small performance boost (around 2ms/frame) by
using a somewhat lower quality downsampling pass. Because N163
generates ultrasonic aliasing, we need to disable its multiplexing
behavior, as otherwise that aliasing sounds awful when run through
a high pass filter that is not tuned to handle it.